Rename Login and SignUp components for better debugging
