+++ src/components/Auth/__tests__/Auth-test.js
@@ -4,8 +4,8 @@ import React from 'react';
 import expect from 'expect';
 import { shallow } from 'enzyme';
 import Auth from '../index.jsx';
-import Login from '../../Login/index.jsx';
-import SignUp from '../../SignUp/index.jsx';
+import Login from '../../Login/Login.jsx';
+import SignUp from '../../SignUp/SignUp.jsx';
 
 describe('Auth', function() {
